% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_forwardoose.R
\name{prophet_forwardoose}
\alias{prophet_forwardoose}
\title{Prophet model fitting with forward out of sample error}
\usage{
prophet_forwardoose(prophet_configuration_object, training_data,
  testing_data, metric = c("ME", "RMSE", "MAE", "MPE", "MAPE", "MASE"))
}
\arguments{
\item{prophet_configuration_object}{The result of your configuration of the prophet_configuration() function template. (Described on the GitHub in the README)}

\item{training_data}{the data upon which the model trains. This must be a dataframe with at least the column ds (standing for date in POSXIT) and y which are the observed values}

\item{testing_data}{the data upon which predictions should be tested. This dataframe should include all the training data. i.e. it should just extend the training data by the values against which you seek to test. Importantly, this must be a dataframe with at least the column ds (standing for date in POSXIT) and y which are the observed values}

\item{metric}{your choice of: "ME", "RMSE", "MAE", "MPE", "MAPE", "MASE". Provide only one option as a character string.}
}
\value{
a list with three named elements: the (1) fitted model, its (2) prediction (over the test data only) and (3) an error aggregation.
}
\description{
Prophet model fitting with forward out of sample error
}
