% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet_rolling_crossvalidation.R
\name{prophet_rolling_crossvalidation}
\alias{prophet_rolling_crossvalidation}
\title{Prophet Rolling Mean Out of Sample Error (Hyndman's Evaluation on a Rolling Forecasting Origin)}
\usage{
prophet_rolling_crossvalidation(prophet_configuration_object, data,
  cores_used, error_metric = c("ME", "RMSE", "MAE", "MPE", "MAPE",
  "MASE"), aggregating_metric = "MEAN", process_starting_row,
  per_validation_period)
}
\arguments{
\item{prophet_configuration_object}{The result of your configuration of the prophet_configuration() function template. (Described on the GitHub in the README)}

\item{data}{a data object containing all data (test and train). This must be a dataframe with at least the column ds (standing for date in POSXIT) and y which are the observed values}

\item{cores_used}{The numebr of cores upon which the paralised optimisation process is allowed to occur. (Leave at least one core open.)}

\item{error_metric}{your choice of: "ME", "RMSE", "MAE", "MPE", "MAPE", "MASE". Provide only one option as a character string.}

\item{aggregating_metric}{either the mean (specify "MEAN") or a vector of weights (which will calculate a weighted mean).}

\item{process_starting_row}{The row for which it is the first point in Hyndman's Evaluation on a rolling forecasting origin}

\item{per_validation_period}{The number of periods per each of the forward Hyndman's Evaluation on a rolling forecasting origin.}
}
\value{
A list with two metrics: Score (which is the value to be maximized) and Pred which is not currely being used.
}
\description{
This function implements Hyndman's Evaluation on a rolling forecasting origin. (See https://robjhyndman.com/hyndsight/tscv/).
The arguments of the function are used to configure it.
}
