% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested_init_points.R
\name{suggested_init_points}
\alias{suggested_init_points}
\title{Title Suggested number of initial points}
\usage{
suggested_init_points(number_of_bounds)
}
\arguments{
\item{number_of_bounds}{The number of bounds passed to the `rBayesianOptimization::BayesianOptimization()` function}
}
\value{
a suggested number of initial points
}
\description{
Title Suggested number of initial points
}
\examples{
suggested_init_points(3)
}
